tests: Factor entity generation into fun
  I want to use this across test modules.
